added tests

+++ src/main/kotlin/com/example/kira/service/SprintService.kt
@@ -3,15 +3,14 @@ package com.example.kira.service
 import com.example.kira.entity.Sprint
 import com.example.kira.dto.SprintCreateRequest
 import com.example.kira.dto.SprintListResponse
+import com.example.kira.exception.SprintNotFoundException
 import com.example.kira.repository.SprintRepository
 import org.springframework.stereotype.Service
 
 @Service
 class SprintService(private val sprintRepository: SprintRepository) {
-    fun getAll(): SprintListResponse {
-        val sprints = sprintRepository.findAll()
-        return SprintListResponse(sprints.count(), sprints.toList())
-    }
+    fun getAll(): SprintListResponse =
+            sprintRepository.findAll().let { SprintListResponse(it.count(), it.toList()) }
 
     fun createSprint(request: SprintCreateRequest): Sprint {
         return sprintRepository.save(
@@ -24,6 +23,6 @@ class SprintService(private val sprintRepository: SprintRepository) {
     }
 
     fun deleteSprintById(id: Long) {
-        sprintRepository.deleteById(id)
+        if (sprintRepository.existsById(id)) sprintRepository.deleteById(id)
     }
 }
