+++ src/main/kotlin/com/example/kira/service/TaskService.kt
@@ -3,15 +3,14 @@ package com.example.kira.service
 import com.example.kira.entity.Task
 import com.example.kira.dto.TaskCreateRequest
 import com.example.kira.dto.TaskListResponse
+import com.example.kira.exception.TaskNotFoundException
 import com.example.kira.repository.TaskRepository
 import org.springframework.stereotype.Service
 
 @Service
 class TaskService(private val taskRepository: TaskRepository) {
-    fun getAll(): TaskListResponse {
-        val tasks = taskRepository.findAll()
-        return TaskListResponse(tasks.count(), tasks.toList())
-    }
+    fun getAll(): TaskListResponse =
+            taskRepository.findAll().let { TaskListResponse(it.count(), it.toList()) }
 
     fun createTask(request: TaskCreateRequest): Task {
         return taskRepository.save(
@@ -25,6 +24,6 @@ class TaskService(private val taskRepository: TaskRepository) {
     }
 
     fun deleteTaskById(id: Long) {
-        taskRepository.deleteById(id)
+        if (taskRepository.existsById(id)) taskRepository.deleteById(id)
     }
 }
